refactor: 어드민 전용 API 및 HTTP 메서드에 일반 유저의 접근을 제한하도록 수정
- /times, /themes, /reservations에 일반 유저는 DELETE할 수 없도록 변경
- /times, /themes에 일반 유저는 POST할 수 없도록 변경
- /admin/reservations 엔드포인트에 일반 유저는 접근할 수 없도록 변경

+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -13,9 +13,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.LoginMember;
 import roomescape.dto.time.AvailableReservationTimeResponse;
 import roomescape.dto.time.ReservationTimeCreateRequest;
 import roomescape.dto.time.ReservationTimeResponse;
+import roomescape.exception.UnauthorizedAccessException;
 import roomescape.service.ReservationTimeService;
 
 @RestController
@@ -41,13 +43,21 @@ public class ReservationTimeController {
     }
 
     @PostMapping
-    public ResponseEntity<ReservationTimeResponse> addReservationTime(@Valid @RequestBody final ReservationTimeCreateRequest requestDto) {
+    public ResponseEntity<ReservationTimeResponse> addReservationTime(@Valid @RequestBody final ReservationTimeCreateRequest requestDto, LoginMember member) {
+        if (member.getRole().equalsIgnoreCase("USER")) {
+            throw new UnauthorizedAccessException("[ERROR] 접근 권한이 없습니다.");
+        }
+
         ReservationTimeResponse responseDto = reservationTimeService.createReservationTime(requestDto);
         return ResponseEntity.created(URI.create("times/" + responseDto.id())).body(responseDto);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteReservationTime(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteReservationTime(@PathVariable Long id, LoginMember member) {
+        if (member.getRole().equalsIgnoreCase("USER")) {
+            throw new UnauthorizedAccessException("[ERROR] 접근 권한이 없습니다.");
+        }
+
         reservationTimeService.deleteReservationTimeById(id);
         return ResponseEntity.noContent().build();
     }
